Feature/version1.0.0

+++ ha-root/src/main/java/jp/co/ha/root/contents/account/response/AccountListApiResponse.java
@@ -56,7 +56,7 @@ public class AccountListApiResponse extends BaseRootApiResponse
 
         /** ユーザID */
         @JsonProperty("seq_user_id")
-        private Integer seqUserId;
+        private Long seqUserId;
         /** メールアドレス */
         @JsonProperty("mail_address")
         private String mailAddress;
@@ -102,7 +102,7 @@ public class AccountListApiResponse extends BaseRootApiResponse
          *
          * @return seqUserId
          */
-        public Integer getSeqUserId() {
+        public Long getSeqUserId() {
             return seqUserId;
         }
 
@@ -112,7 +112,7 @@ public class AccountListApiResponse extends BaseRootApiResponse
          * @param seqUserId
          *     ユーザID
          */
-        public void setSeqUserId(Integer seqUserId) {
+        public void setSeqUserId(Long seqUserId) {
             this.seqUserId = seqUserId;
         }
